refactor: merge orchestration_state + orchestration_store into unified StateStore
Wave A piece 1 of the goldfive refactor. The split between
``orchestration_state`` (module-level write helpers + key constants) and
``orchestration_store`` (typed ``OrchestrationStore`` read handle plus a
handful of writes) was a half-finished Phase-1 stage — both modules
touched the same ADK session dict via different APIs. Collapses both
onto a single ``goldfive/state_store.py`` module exposing:

* All ``KEY_*`` / ``LIFECYCLE_*`` constants previously in
  ``orchestration_state``.
* All free helpers (``write``, ``read``, ``set_current_plan``,
  ``set_current_task``, ``sync_current_task_from_transition``,
  ``rotate_current_task_id``, drift lifecycle helpers, …) preserved
  verbatim — call sites that imported via
  ``from goldfive import orchestration_state as _ostate`` only need the
  one import line updated.
* The ``ActiveSteer`` / ``DelegationPin`` / ``ReasoningBinding``
  dataclasses, the ``BindingSource`` enum, and the typed handle —
  renamed ``OrchestrationStore`` -> ``StateStore`` — with reads and
  writes on one object.

Pure mechanical refactor — no behaviour change. State-ownership audit
catalog unchanged (the catalog tracks ADK-state writers, not goldfive
``Session.state`` writers).

* Old modules deleted; no back-compat shim — single-line import update
  + ``OrchestrationStore`` -> ``StateStore`` rename across 37 files.
* New ``tests/test_state_store.py`` covers the merged surface against
  a plain dict (no ``Session`` / ADK runtime needed).
* Full test suite green (2459 passed, 59 skipped).
