--- goldfive/orchestration_store.py
@@ -1,970 +1,0 @@
-"""Typed accessor over goldfive ``Session.state`` for orchestration data.
-
-Phase 1 of goldfive#271 — see ``docs/design/STATE-OWNERSHIP-CONTRACT.md``
-for the contract this layer enforces, and Phase 0's audit catalog
-(``§5`` of that doc) for the full set of writers/readers Phase 2 will
-collapse onto this surface.
-
-What this module is
--------------------
-
-A thin, typed handle wrapping goldfive's own
-:class:`~goldfive.types.Session.state` dict. Replaces ad-hoc
-``session.state.get('goldfive.foo')`` reads and ``session.state[...] = v``
-writes scattered across the codebase with named methods + small typed
-result objects (``ActiveSteer``, ``DelegationPin``, ``ReasoningBinding``).
-
-What this module is NOT
------------------------
-
-* **NOT a wrapper over ADK's ``session.state``.** That dict is owned by
-  ADK; goldfive callbacks must not mutate it (Phase 0 contract §3.3).
-  The ADK-side reads in
-  :mod:`~goldfive.adapters._adk_dynainst` and
-  :mod:`~goldfive.planners.goldfive_planner` consult ADK's state
-  copy that the bridge (Phase 2 migration target V2) maintains; this
-  store is the goldfive-internal source of truth that bridge reads
-  *from*.
-* **NOT a replacement for** :mod:`goldfive.orchestration_state`'s
-  primitive read/write helpers. Those keep their existing call sites
-  untouched in Phase 1 — this module is a read-side veneer that
-  supersedes the *ad-hoc* ``state.get(...)`` callers cataloged in the
-  Phase 0 audit. Phase 2 collapses the writers; Phase 3 collapses the
-  bridge.
-* **NOT exhaustive.** Phase 1's surface covers only the recurring
-  operations Phase 1's brief calls out plus the new reasoning-extracted
-  binding. Subsequent phases extend the surface as they migrate
-  additional sites.
-
-Initial surface (Phase 1)
--------------------------
-
-Reads (replacing ad-hoc ``state.get('goldfive.xxx')`` callers):
-
-* :meth:`OrchestrationStore.pin_current_task` — current task pin id
-* :meth:`OrchestrationStore.pin_current_task_revision` — pin revision
-* :meth:`OrchestrationStore.get_active_steer` — active steer, typed
-* :meth:`OrchestrationStore.get_correction` — pending-correction body
-* :meth:`OrchestrationStore.get_pending_delegation` — per-fc_id pin
-
-Writes:
-
-* :meth:`OrchestrationStore.set_pin_current_task` — write the pin
-  (with a documented :class:`BindingSource`)
-* :meth:`OrchestrationStore.record_reasoning_extracted_binding` —
-  the **only NEW write site** introduced by Phase 1. Records the LLM
-  judge's stated-intent binding so the pin-resolution ladder's signal
-  6 can consume it as a real signal.
-* :meth:`OrchestrationStore.clear_reasoning_extracted_binding` —
-  companion clear (called on task transition).
-
-Reasoning-extracted bindings
-----------------------------
-
-A new orchestration-state slot,
-``goldfive.reasoning_extracted_bindings`` (a dict keyed by agent name),
-records the LLM judge's stated-intent attribution from
-:func:`~goldfive.drift.reasoning_judge.classify_reasoning_drift`. When
-the judge returns ``focused_task_id`` + ``focus_confidence`` and the
-confidence is above a configured threshold, the steerer records a
-binding here; the pin-resolution ladder's signal 6 consults it.
-
-Threshold semantics:
-
-* The threshold is a steerer-level config knob (default ``0.7``); the
-  store stamps whatever confidence it's given and lets the consumer
-  decide.
-* The recorded binding includes the originating ``run_id`` /
-  ``session_id`` / ``agent_name`` / ``task_id`` / ``confidence``
-  plus a ``recorded_at_turn`` so consumers can dismiss stale bindings.
-
-Surface evolution
------------------
-
-Each Phase-2 migration of a writer adds a corresponding ``set_*`` /
-``clear_*`` method here. As the bridge collapses (Phase 3) the read
-side migrates to take a ``ReadonlyContext`` adapter; until then the
-reads return cached views of goldfive ``Session.state`` and the bridge
-copies them onto ADK's state.
-"""
-
-from __future__ import annotations
-
-import asyncio
-import dataclasses
-import threading
-from collections.abc import Mapping, MutableMapping
-from enum import StrEnum
-from typing import TYPE_CHECKING, Any
-
-from goldfive import orchestration_state as _ostate
-from goldfive.adapters import _adk_dynainst as _dynainst
-
-if TYPE_CHECKING:  # pragma: no cover — type-check only
-    from goldfive.types import Session
-
-
-__all__ = [
-    "ActiveSteer",
-    "BindingSource",
-    "DelegationPin",
-    "OrchestrationStore",
-    "PENDING_DELEGATIONS_KEY",
-    "REASONING_BINDINGS_KEY",
-    "ReasoningBinding",
-]
-
-
-# ---------------------------------------------------------------------------
-# Active-invocation registry (goldfive#271 Phase 3.5 component 1)
-# ---------------------------------------------------------------------------
-#
-# Per-session asyncio.Task registry keyed by ``invocation_id``. Owned by
-# OrchestrationStore (the per-session orchestration-state surface) rather
-# than the ADK plugin instance, restoring the Phase 0 state-ownership
-# contract: per-session orchestration data lives on the store, not on
-# adapter plugin instances.
-#
-# Why a module-level dict rather than a goldfive ``Session.state`` slot:
-# :class:`asyncio.Task` is intentionally non-serializable (it holds a
-# reference to a live coroutine + the running loop), and
-# :mod:`goldfive.orchestration_state.write` enforces "values must
-# round-trip cleanly through sinks". Storing tasks under the
-# ``goldfive.*`` prefix would either break that invariant or force every
-# sink to special-case the slot. Instead we keep tasks in this
-# module-level registry and key the outer dict by ``Session.id`` so the
-# OrchestrationStore-as-view contract still holds: each store instance
-# only sees its own session's tasks.
-#
-# Lock: a process-wide :class:`threading.Lock` protects the outer dict's
-# structural mutations (sub-dict insert / pop). Inner-dict reads /
-# writes are single-threaded per session because ADK callbacks run on
-# one event loop, but a concurrent ``request_invocation_cancel`` from a
-# different session could otherwise race the structural setdefault.
-_ACTIVE_INVOCATION_TASKS: dict[str, dict[str, asyncio.Task[Any]]] = {}
-_ACTIVE_INVOCATION_LOCK = threading.Lock()
-
-# Companion registry: per-session set of invocation ids for which a
-# cooperative cancel has been requested (goldfive#242). Stamped
-# synchronously at the top of
-# :meth:`~goldfive.steerer.DefaultSteerer.request_invocation_cancel`
-# so the late-drift gate can short-circuit drifts that fire during the
-# 4-8s window between the cancel-request landing and ADK winding down
-# the cancelled invocations (during which ``active_invocation_ids()``
-# still lists them).
-#
-# Same locking discipline as the active-task registry above. Cleared by
-# :meth:`OrchestrationStore.clear_active_invocations` so a session
-# teardown wipes both registries in one shot.
-_CANCEL_REQUESTED_INVOCATIONS: dict[str, set[str]] = {}
-
-
-# State key for the new reasoning-extracted bindings slot. Lives under
-# the goldfive prefix so :func:`goldfive.orchestration_state.write`
-# accepts it. Value shape: ``dict[agent_name, ReasoningBinding-as-dict]``
-# for cheap JSON serialisation by sinks that round-trip the state dict.
-REASONING_BINDINGS_KEY = "goldfive.reasoning_extracted_bindings"
-
-# State key for the per-``function_call_id`` delegation-pin map
-# (goldfive#241 Item 3-bis, V4 in the Phase 0 audit catalog). Phase 2.1
-# of goldfive#271 — consolidated here so writers and readers agree on
-# one source of truth on goldfive's own ``Session.state``.
-#
-# Value shape: ``dict[function_call_id, {task_id, revision, tool_args?}]``.
-# Legacy entries stamped by pre-#266 adapters were bare strings; the
-# :class:`DelegationPin` accessor normalises both shapes.
-PENDING_DELEGATIONS_KEY = "goldfive.pending_delegations"
-
-
-# ---------------------------------------------------------------------------
-# Typed result objects
-# ---------------------------------------------------------------------------
-
-
-class BindingSource(StrEnum):
-    """Origin of a current-task pin write.
-
-    Stamped onto the pin so the pin-resolution ladder's events can
-    distinguish a pin set by an agent-turn callback from one set by
-    a delegation site, a steerer rotation, or a reasoning-extracted
-    binding.
-
-    Phase 1 only consumes the value for observability + the new
-    reasoning-extracted-binding signal in the pin ladder; Phase 2's
-    full migration will use it to attribute every catalogued writer.
-    """
-
-    DELEGATION_PIN = "delegation_pin"
-    AGENT_CALLBACK = "agent_callback"
-    REASONING = "reasoning"
-    CORRECTION_TARGET = "correction_target"
-    STEERER_ROTATION = "steerer_rotation"
-    LOW_CONFIDENCE = "low_confidence"
-    UNKNOWN = "unknown"
-
-
-@dataclasses.dataclass(frozen=True)
-class ActiveSteer:
-    """Typed view of the active steer slots on goldfive ``Session.state``.
-
-    Wraps the four
-    :data:`~goldfive.orchestration_state.KEY_ACTIVE_STEER_BODY` /
-    ``_AT_TURN`` / ``_AUTHOR`` / ``_SOURCE`` keys so callers don't
-    string-fish through the state dict.
-    """
-
-    body: str
-    at_turn: int
-    author: str
-    source: str  # "user" | "goldfive" | ""
-
-    def is_active(self) -> bool:
-        """True when a steer is currently set (non-empty body)."""
-        return bool(self.body)
-
-
-@dataclasses.dataclass(frozen=True)
-class DelegationPin:
-    """Typed view of a single ``goldfive.pending_delegations`` entry.
-
-    The on-disk shape supports both legacy bare-string and the
-    versioned ``{task_id, revision, tool_args}`` dict (goldfive#266 /
-    F7). This dataclass normalises both shapes for callers.
-    """
-
-    task_id: str
-    revision: int = 0
-    tool_args: Mapping[str, Any] | None = None
-
-    def is_set(self) -> bool:
-        return bool(self.task_id)
-
-
-@dataclasses.dataclass(frozen=True)
-class ReasoningBinding:
-    """Typed view of a reasoning-extracted binding.
-
-    Persisted as a plain dict under
-    :data:`REASONING_BINDINGS_KEY[agent_name]` so sinks can round-trip
-    the state dict; this dataclass is the in-process view.
-
-    ``recorded_at_turn`` is the session sequence value at the moment
-    the binding was recorded; consumers can compare it against the
-    current sequence to dismiss bindings older than N turns. ``0``
-    means "no sequence recorded" (e.g. test fixture without a session
-    sequence counter).
-    """
-
-    agent_name: str
-    task_id: str
-    confidence: float
-    recorded_at_turn: int = 0
-    run_id: str = ""
-    session_id: str = ""
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "agent_name": self.agent_name,
-            "task_id": self.task_id,
-            "confidence": float(self.confidence),
-            "recorded_at_turn": int(self.recorded_at_turn),
-            "run_id": self.run_id,
-            "session_id": self.session_id,
-        }
-
-    @classmethod
-    def from_dict(cls, raw: Any) -> ReasoningBinding | None:
-        """Construct from a state-dict value, or ``None`` on garbage.
-
-        Tolerant of partial dicts (sink-shaped, persistence-restored)
-        so a missing ``run_id`` / ``session_id`` doesn't drop a
-        legitimate binding.
-        """
-        if not isinstance(raw, Mapping):
-            return None
-        try:
-            task_id = str(raw.get("task_id", "") or "")
-            confidence = float(raw.get("confidence", 0.0) or 0.0)
-            agent_name = str(raw.get("agent_name", "") or "")
-        except (TypeError, ValueError):
-            return None
-        if not task_id:
-            return None
-        try:
-            recorded_at_turn = int(raw.get("recorded_at_turn", 0) or 0)
-        except (TypeError, ValueError):
-            recorded_at_turn = 0
-        return cls(
-            agent_name=agent_name,
-            task_id=task_id,
-            confidence=confidence,
-            recorded_at_turn=recorded_at_turn,
-            run_id=str(raw.get("run_id", "") or ""),
-            session_id=str(raw.get("session_id", "") or ""),
-        )
-
-
-# ---------------------------------------------------------------------------
-# OrchestrationStore — typed handle
-# ---------------------------------------------------------------------------
-
-
-class OrchestrationStore:
-    """Typed handle over goldfive ``Session.state``.
-
-    Construct with :meth:`for_session` (when you have a goldfive
-    :class:`~goldfive.types.Session`) or :meth:`for_state` (when you
-    only have the state dict, e.g. in test scaffolding or in callbacks
-    that reach the goldfive Session via
-    :data:`~goldfive.adapters._adk_plugin.SESSION_CONTEXT_STATE_KEY`).
-
-    All accessors are forgiving: missing / malformed entries return
-    typed defaults rather than raising. The store is a *view* — it
-    does not own the underlying dict and concurrent writers can mutate
-    it between calls. Callers who need a snapshot should call once and
-    cache.
-    """
-
-    __slots__ = ("_state", "_session_id")
-
-    def __init__(self, state: Any, *, session_id: str = "") -> None:
-        # We accept any mapping-shaped object — goldfive's Session.state
-        # is a plain dict; tests sometimes pass a ``MappingProxyType``
-        # over the same dict. Read paths tolerate both; write paths
-        # require :class:`MutableMapping` (raised lazily by the helper).
-        self._state = state if isinstance(state, Mapping) else {}
-        # ``session_id`` keys the active-invocation task registry below.
-        # Empty when the store was built for an arbitrary state dict
-        # (test scaffolding, callbacks reaching the store with no
-        # Session in scope) — those callers cannot drive the
-        # active-invocation registry, but every other read/write path
-        # remains usable.
-        self._session_id = str(session_id or "")
-
-    # -- Constructors ----------------------------------------------------
-
-    @classmethod
-    def for_session(cls, session: Session | None) -> OrchestrationStore:
-        """Build a store backed by ``session.state``.
-
-        ``None`` yields an empty-state store (writes are silently
-        dropped) so callers who can't guarantee a session — e.g.
-        defensive paths inside ADK callbacks — never raise.
-        """
-        if session is None:
-            return cls({})
-        return cls(
-            getattr(session, "state", {}),
-            session_id=str(getattr(session, "id", "") or ""),
-        )
-
-    @classmethod
-    def for_state(cls, state: Any, *, session_id: str = "") -> OrchestrationStore:
-        """Build a store backed by an arbitrary state dict."""
-        return cls(state, session_id=session_id)
-
-    # -- Internal helpers -----------------------------------------------
-
-    def _get(self, key: str, default: Any = None) -> Any:
-        return _ostate.read(self._state, key, default)
-
-    # -- Read: pin -------------------------------------------------------
-
-    def pin_current_task(self) -> str:
-        """Return ``goldfive.current_task_id``, or ``""`` when unset.
-
-        Replaces ``state.get('goldfive.current_task_id', '')`` with
-        a typed accessor. Phase 1's primary read-migration target on
-        the goldfive side.
-        """
-        value = self._get(_ostate.KEY_CURRENT_TASK_ID, "")
-        if isinstance(value, str):
-            return value
-        return ""
-
-    def pin_current_task_title(self) -> str:
-        """Return ``goldfive.current_task_title``, or ``""``."""
-        value = self._get(_ostate.KEY_CURRENT_TASK_TITLE, "")
-        if isinstance(value, str):
-            return value
-        return ""
-
-    def pin_current_task_revision(self) -> int:
-        """Return the revision stamp on the current pin (0 when unset)."""
-        return _ostate.read_current_task_revision(self._state)
-
-    # -- Write: pin ------------------------------------------------------
-
-    def set_pin_current_task(
-        self,
-        task_id: str,
-        *,
-        source: BindingSource = BindingSource.UNKNOWN,
-        revision: int | None = None,
-        title: str = "",
-    ) -> None:
-        """Stamp the current-task pin.
-
-        ``source`` is the :class:`BindingSource` documenting which
-        ladder-rung / callback wrote the pin. Phase 1 uses it for
-        observability only — Phase 2 wires it through every catalogued
-        writer for full attribution. Passing ``BindingSource.UNKNOWN``
-        is fine for code paths whose attribution hasn't been migrated
-        yet.
-
-        ``revision`` (when provided) stamps
-        ``goldfive.current_task_revision`` alongside the id.
-        ``None`` leaves the existing revision untouched.
-
-        No-op when ``task_id`` is empty — callers should use
-        :meth:`clear_pin_current_task` to clear.
-        """
-        if not task_id:
-            return
-        if not isinstance(self._state, dict):
-            # Read-only state (e.g. a MappingProxyType view); silently
-            # drop the write so the caller's defensive path remains
-            # safe. Production state is always a mutable dict.
-            return
-        # Reuse the orchestration_state primitives so the
-        # ``goldfive.*``-prefix assertion still fires; this keeps the
-        # store's writes funneled through the same place the catalog
-        # is verified against.
-        _ostate.write(self._state, _ostate.KEY_CURRENT_TASK_ID, str(task_id))
-        if title:
-            _ostate.write(self._state, _ostate.KEY_CURRENT_TASK_TITLE, str(title))
-        if revision is not None:
-            _ostate.stamp_current_task_revision(self._state, int(revision))
-        # ``source`` is recorded as part of the binding registry below
-        # rather than on the pin slot itself; the pin slot pre-dates
-        # the source vocabulary and Phase 2 will collapse the two.
-        _ = source  # documented for migration; not stored on the pin slot
-
-    def clear_pin_current_task(self) -> None:
-        """Clear the current-task pin slots. Idempotent."""
-        if not isinstance(self._state, dict):
-            return
-        _ostate.clear_current_task(self._state)
-
-    # -- Read: active steer ---------------------------------------------
-
-    def get_active_steer(self) -> ActiveSteer | None:
-        """Return the active steer, or ``None`` when no steer is set.
-
-        Replaces the four-key dance
-        ``state.get(KEY_ACTIVE_STEER_BODY, '')`` etc. with one typed
-        read. ``None`` is returned when the body is empty (the canonical
-        "no steer" signal) so callers can ``if store.get_active_steer():``
-        without re-checking ``.body``.
-        """
-        body = self._get(_ostate.KEY_ACTIVE_STEER_BODY, "")
-        if not isinstance(body, str) or not body:
-            return None
-        try:
-            at_turn = int(self._get(_ostate.KEY_ACTIVE_STEER_AT_TURN, 0) or 0)
-        except (TypeError, ValueError):
-            at_turn = 0
-        author_raw = self._get(_ostate.KEY_ACTIVE_STEER_AUTHOR, "")
-        source_raw = self._get(_ostate.KEY_ACTIVE_STEER_SOURCE, "")
-        return ActiveSteer(
-            body=body,
-            at_turn=at_turn,
-            author=str(author_raw or ""),
-            source=str(source_raw or ""),
-        )
-
-    # -- Read: goals summary --------------------------------------------
-
-    def goals_summary(self) -> str:
-        """Return ``goldfive.goals_summary``, or ``""`` when unset.
-
-        Pre-formatted comma-joined string maintained by
-        :func:`goldfive.orchestration_state.refresh_goals_summary`. The
-        planner consumes this value verbatim for its per-turn
-        instruction block.
-        """
-        value = self._get(_ostate.KEY_GOALS_SUMMARY, "")
-        if isinstance(value, str):
-            return value
-        return ""
-
-    # -- Read: cancelled function-call ids -------------------------------
-
-    def cancelled_function_call_ids(self) -> list[str]:
-        """Return the list of cancelled ``function_call`` ids.
-
-        Reuses :func:`goldfive.orchestration_state.read_cancelled_function_call_ids`
-        so the list-shape guard (non-list -> ``[]``) is centralised.
-        """
-        return _ostate.read_cancelled_function_call_ids(self._state)
-
-    # -- Read: correction ------------------------------------------------
-
-    def get_correction(self, agent_name: str, task_id: str) -> Any:
-        """Return the pending-correction value for an ``(agent, task)``.
-
-        The shape is whatever
-        :mod:`goldfive._correction_injection` writes — typically a
-        :class:`Mapping` with ``superseded_task_id`` /
-        ``revision_number`` / etc. (rendered via
-        :func:`goldfive.adapters._adk_dynainst.format_correction_block`)
-        but tests / external callers may have written a pre-rendered
-        string. ``None`` means no pending correction.
-        """
-        if not agent_name or not task_id:
-            return None
-        key = _dynainst.pending_correction_key(agent_name, task_id)
-        return self._get(key, None)
-
-    def has_correction(self, agent_name: str, task_id: str) -> bool:
-        """True when a pending correction exists for ``(agent, task)``."""
-        return self.get_correction(agent_name, task_id) is not None
-
-    def iter_corrections_for_agent(self, agent_name: str) -> list[str]:
-        """Return every task_id with a pending correction for ``agent_name``.
-
-        Used by pin-ladder signal 6 to enumerate correction targets
-        without rebuilding the prefix-matching loop at every call site.
-        """
-        if not agent_name or not isinstance(self._state, Mapping):
-            return []
-        # Strip a compound prefix so callers passing the bare or the
-        # compound form both find the writer's bare-form keys. Mirrors
-        # the matching the existing ``_task_from_pending_correction``
-        # helper does inline today.
-        bare = agent_name.rsplit(":", 1)[-1]
-        prefix = f"goldfive.pending_corrections.{bare}."
-        out: list[str] = []
-        for key in self._state:
-            if not isinstance(key, str):
-                continue
-            if not key.startswith(prefix):
-                continue
-            tid = key[len(prefix) :]
-            if tid:
-                out.append(tid)
-        return out
-
-    # -- Read: pending delegations --------------------------------------
-
-    def get_pending_delegation(self, fc_id: str) -> DelegationPin | None:
-        """Return the per-``function_call_id`` delegation pin, or ``None``.
-
-        Tolerant of both legacy bare-string entries and the versioned
-        ``{task_id, revision, tool_args}`` dict shape (goldfive#266 +
-        F7). Callers used to inline the shape-test; this normalises.
-        """
-        if not fc_id:
-            return None
-        pend = self._get(PENDING_DELEGATIONS_KEY, None)
-        if not isinstance(pend, Mapping):
-            return None
-        raw = pend.get(fc_id)
-        if raw is None:
-            return None
-        if isinstance(raw, str):
-            tid = raw.strip()
-            if not tid:
-                return None
-            return DelegationPin(task_id=tid)
-        if isinstance(raw, Mapping):
-            tid = str(raw.get("task_id", "") or "").strip()
-            if not tid:
-                return None
-            try:
-                rev = int(raw.get("revision", 0) or 0)
-            except (TypeError, ValueError):
-                rev = 0
-            args = raw.get("tool_args")
-            if not isinstance(args, Mapping):
-                args = None
-            return DelegationPin(task_id=tid, revision=rev, tool_args=args)
-        return None
-
-    def iter_pending_delegations(self) -> Mapping[str, Any]:
-        """Return the raw ``pending_delegations`` map (or empty mapping).
-
-        The plugin's pin-resolution ladder needs to walk every entry
-        (signal 1 iterates to find a task-id match across all parallel
-        dispatches; signal 3 merges every entry's ``tool_args`` into a
-        single token bag for scoring). Returns the live dict so callers
-        can iterate with ``.values()`` / ``.items()`` without a copy.
-        """
-        pend = self._get(PENDING_DELEGATIONS_KEY, None)
-        if isinstance(pend, Mapping):
-            return pend
-        return {}
-
-    # -- Write: pending delegations -------------------------------------
-
-    def set_pending_delegation(
-        self,
-        fc_id: str,
-        *,
-        task_id: str,
-        revision: int = 0,
-        tool_args: Mapping[str, Any] | None = None,
-    ) -> None:
-        """Stamp a per-``function_call_id`` delegation pin.
-
-        V4 of the Phase 0 audit (goldfive#271) — every delegation-site
-        write now lands here on goldfive ``Session.state`` rather than
-        on ADK ``session.state``. The pin-resolution ladder + reporting
-        handlers consult this same store, so a single write is enough.
-
-        ``tool_args`` (when provided as a non-empty Mapping) is stamped
-        alongside the pin so signal 3 of the ladder can score
-        candidates against the parent's literal dispatch args (F7 /
-        #265 followup). Empty / non-mapping args are dropped — the
-        scorer treats those as zero-token signals.
-
-        No-op when ``fc_id`` or ``task_id`` is empty, or when the
-        backing state is not a mutable dict (defensive — production
-        state is always a dict; tests sometimes pass MappingProxyType
-        snapshots).
-        """
-        if not fc_id or not task_id:
-            return
-        if not isinstance(self._state, dict):
-            return
-        existing = self._state.get(PENDING_DELEGATIONS_KEY)
-        bucket: dict[str, Any]
-        if isinstance(existing, dict):
-            bucket = existing
-        else:
-            bucket = {}
-        entry: dict[str, Any] = {
-            "task_id": str(task_id),
-            "revision": int(revision),
-        }
-        if isinstance(tool_args, Mapping) and tool_args:
-            entry["tool_args"] = dict(tool_args)
-        bucket[str(fc_id)] = entry
-        _ostate.write(self._state, PENDING_DELEGATIONS_KEY, bucket)
-
-    # -- Read: reasoning-extracted bindings (NEW Phase 1) ---------------
-
-    def get_reasoning_extracted_binding(
-        self,
-        agent_name: str,
-    ) -> ReasoningBinding | None:
-        """Return the most recent reasoning-extracted binding for ``agent_name``.
-
-        Phase 1's NEW read path — consumed by the pin-resolution
-        ladder's signal 6 (and by future correction / drift logic that
-        wants to consult the LLM judge's stated-intent attribution).
-
-        Returns ``None`` when no binding exists for the agent (or when
-        the recorded entry is malformed). Callers gate on confidence
-        themselves; the store stamps whatever the writer recorded.
-
-        The lookup strips compound-form prefixes the same way
-        :meth:`iter_corrections_for_agent` does, so a compound
-        ``"client42:agent_x"`` finds the bare ``agent_x`` binding the
-        judge recorded.
-        """
-        if not agent_name:
-            return None
-        registry = self._get(REASONING_BINDINGS_KEY, None)
-        if not isinstance(registry, Mapping):
-            return None
-        # Try the exact form, then the bare-form fallback so
-        # compound-named callers (``"client:foo"``) still match a bare
-        # binding the judge recorded for ``foo``.
-        raw = registry.get(agent_name)
-        if raw is None:
-            bare = agent_name.rsplit(":", 1)[-1]
-            if bare and bare != agent_name:
-                raw = registry.get(bare)
-        if raw is None:
-            return None
-        return ReasoningBinding.from_dict(raw)
-
-    # -- Write: reasoning-extracted bindings (NEW Phase 1) --------------
-
-    def record_reasoning_extracted_binding(
-        self,
-        *,
-        agent_name: str,
-        task_id: str,
-        confidence: float,
-        recorded_at_turn: int = 0,
-        run_id: str = "",
-        session_id: str = "",
-    ) -> ReasoningBinding | None:
-        """Stamp a reasoning-extracted binding for ``agent_name``.
-
-        Phase 1's only NEW write site. Called by the steerer's reasoning
-        observation path when
-        :func:`~goldfive.drift.reasoning_judge.classify_reasoning_drift`
-        returns a ``focused_task_id`` with confidence above the
-        configured threshold.
-
-        Returns the recorded :class:`ReasoningBinding` (for caller
-        observability) or ``None`` when the inputs were rejected
-        (empty ``agent_name`` / ``task_id``, or read-only state).
-        Confidence is clamped to ``[0.0, 1.0]``.
-        """
-        if not agent_name or not task_id:
-            return None
-        if not isinstance(self._state, dict):
-            return None
-        try:
-            conf = float(confidence)
-        except (TypeError, ValueError):
-            conf = 0.0
-        conf = max(0.0, min(1.0, conf))
-        binding = ReasoningBinding(
-            agent_name=str(agent_name),
-            task_id=str(task_id),
-            confidence=conf,
-            recorded_at_turn=int(recorded_at_turn),
-            run_id=str(run_id or ""),
-            session_id=str(session_id or ""),
-        )
-        # Read-modify-write the registry so unrelated agents' bindings
-        # are preserved. Tolerate a non-dict existing value (legacy
-        # state shape) by replacing it with a fresh registry rather
-        # than appending to a malformed structure.
-        registry = self._state.get(REASONING_BINDINGS_KEY)
-        if not isinstance(registry, dict):
-            registry = {}
-        # Stamp under the bare form so :meth:`get_reasoning_extracted_binding`
-        # finds it for both compound and bare lookups. The judge owns
-        # the agent-name normalisation policy; the store records the
-        # form the caller supplied (callers normalise before calling).
-        registry[binding.agent_name] = binding.to_dict()
-        _ostate.write(self._state, REASONING_BINDINGS_KEY, registry)
-        return binding
-
-    def clear_reasoning_extracted_binding(self, agent_name: str) -> None:
-        """Drop the binding for ``agent_name`` (idempotent).
-
-        Called on task transition for the agent so a binding that
-        targeted the now-completed task doesn't leak forward and
-        mis-pin the next invocation.
-        """
-        if not agent_name or not isinstance(self._state, dict):
-            return
-        registry = self._state.get(REASONING_BINDINGS_KEY)
-        if not isinstance(registry, dict):
-            return
-        registry.pop(agent_name, None)
-        # Also try the bare form in case the writer used compound but
-        # the caller is clearing with the bare form (or vice-versa).
-        bare = agent_name.rsplit(":", 1)[-1]
-        if bare and bare != agent_name:
-            registry.pop(bare, None)
-        _ostate.write(self._state, REASONING_BINDINGS_KEY, registry)
-
-    # -- Active-invocation registry (goldfive#271 Phase 3.5 component 1) ---
-    #
-    # Per-session ``invocation_id -> asyncio.Task`` map. The goldfive task
-    # boundary (the wrapper around each agent invocation in the ADK
-    # plugin's ``before_agent_callback`` / ``after_agent_callback``
-    # try/finally arc) registers the running task here on entry and
-    # deregisters on exit. The steerer's
-    # :meth:`request_invocation_cancel` looks up the task to fire
-    # ``task.cancel()`` on the live invocation.
-    #
-    # Migrated off ``_GoldfiveADKPlugin._invocation_tasks`` (PR #303 →
-    # Phase 3.5 #305): per-session orchestration state belongs on the
-    # store, not on the adapter plugin instance. Restores the Phase 0
-    # state-ownership contract.
-
-    def register_invocation_task(
-        self,
-        invocation_id: str,
-        task: asyncio.Task[Any],
-    ) -> None:
-        """Register the running asyncio.Task driving ``invocation_id``.
-
-        Called from the goldfive boundary wrapper at entry. No-op when
-        ``invocation_id`` is empty or the store has no session id (e.g.
-        tests that construct a bare-state store).
-        """
-        if not invocation_id or not self._session_id:
-            return
-        with _ACTIVE_INVOCATION_LOCK:
-            bucket = _ACTIVE_INVOCATION_TASKS.setdefault(self._session_id, {})
-        bucket[str(invocation_id)] = task
-
-    def deregister_invocation_task(self, invocation_id: str) -> None:
-        """Drop ``invocation_id`` from the registry. Idempotent.
-
-        Called from the goldfive boundary wrapper's ``finally`` clause
-        at exit. The bucket itself is left in place; cleanup of the
-        outer entry happens in :meth:`clear_active_invocations`.
-        """
-        if not invocation_id or not self._session_id:
-            return
-        with _ACTIVE_INVOCATION_LOCK:
-            bucket = _ACTIVE_INVOCATION_TASKS.get(self._session_id)
-        if bucket is None:
-            return
-        bucket.pop(str(invocation_id), None)
-
-    def get_invocation_task(self, invocation_id: str) -> asyncio.Task[Any] | None:
-        """Return the registered task for ``invocation_id``, or ``None``."""
-        if not invocation_id or not self._session_id:
-            return None
-        with _ACTIVE_INVOCATION_LOCK:
-            bucket = _ACTIVE_INVOCATION_TASKS.get(self._session_id)
-        if bucket is None:
-            return None
-        return bucket.get(str(invocation_id))
-
-    def active_invocation_ids(self) -> list[str]:
-        """Return the ids of every currently-registered invocation.
-
-        Diagnostic / test helper. Empty list when no session id or no
-        registered invocations.
-        """
-        if not self._session_id:
-            return []
-        with _ACTIVE_INVOCATION_LOCK:
-            bucket = _ACTIVE_INVOCATION_TASKS.get(self._session_id)
-        if bucket is None:
-            return []
-        return list(bucket.keys())
-
-    def clear_active_invocations(self) -> None:
-        """Drop every registered task for this session. Idempotent.
-
-        Called from the adapter's dispatch teardown so a stale handle
-        cannot target the next invocation. Also clears the companion
-        cancel-requested registry (goldfive#242) so a fresh dispatch
-        on the same session id starts with a clean slate.
-        """
-        if not self._session_id:
-            return
-        with _ACTIVE_INVOCATION_LOCK:
-            _ACTIVE_INVOCATION_TASKS.pop(self._session_id, None)
-            _CANCEL_REQUESTED_INVOCATIONS.pop(self._session_id, None)
-
-    # -- Cancel-requested registry (goldfive#242) -----------------------
-    #
-    # Closes the iter-11D race between
-    # :meth:`~goldfive.steerer.DefaultSteerer.request_invocation_cancel`
-    # (synchronous flag flip) and
-    # :meth:`active_invocation_ids` (transitions to empty only AFTER
-    # ADK winds down each cancelled invocation, ~4-8s later). The
-    # late-drift gate consults this set; a drift that fires during that
-    # window sees ``cancel_requested_invocation_ids()`` non-empty and
-    # is treated as late, even though the active-task registry still
-    # lists the cancelled invocation.
-
-    def mark_invocation_cancel_requested(self, invocation_id: str) -> None:
-        """Stamp ``invocation_id`` as having a pending cancel request.
-
-        Called synchronously from the top of
-        :meth:`~goldfive.steerer.DefaultSteerer.request_invocation_cancel`
-        before any plugin / async work. Idempotent; multiple cancel
-        requests for the same id collapse to one entry.
-
-        No-op when ``invocation_id`` is empty or the store has no
-        session id (e.g. tests that construct a bare-state store).
-        """
-        if not invocation_id or not self._session_id:
-            return
-        with _ACTIVE_INVOCATION_LOCK:
-            bucket = _CANCEL_REQUESTED_INVOCATIONS.setdefault(self._session_id, set())
-            bucket.add(str(invocation_id))
-
-    def is_invocation_cancel_requested(self, invocation_id: str) -> bool:
-        """Return True iff a cancel was requested for ``invocation_id``."""
-        if not invocation_id or not self._session_id:
-            return False
-        with _ACTIVE_INVOCATION_LOCK:
-            bucket = _CANCEL_REQUESTED_INVOCATIONS.get(self._session_id)
-        if bucket is None:
-            return False
-        return str(invocation_id) in bucket
-
-    def cancel_requested_invocation_ids(self) -> list[str]:
-        """Return the ids of every invocation with a pending cancel.
-
-        Empty list when no session id or no pending cancels.
-        """
-        if not self._session_id:
-            return []
-        with _ACTIVE_INVOCATION_LOCK:
-            bucket = _CANCEL_REQUESTED_INVOCATIONS.get(self._session_id)
-        if bucket is None:
-            return []
-        return list(bucket)
-
-    # -- Active drift conditions (goldfive#271 PR1) ---------------------
-
-    def active_drifts(self) -> list[_ostate.Drift]:
-        """Return all in-flight drift conditions on this session.
-
-        Backed by the ``goldfive.active_drifts`` slot on
-        ``session.state``. Each entry is a :class:`Drift` snapshot;
-        the list is freshly materialised from the underlying dict so
-        callers can mutate the results without affecting subsequent
-        reads.
-        """
-        return _ostate.list_active_drifts(self._state)
-
-    def get_active_drift(self, condition_id: str) -> _ostate.Drift | None:
-        """Look up an in-flight condition by id, or ``None`` when absent."""
-        return _ostate.get_active_drift(self._state, condition_id)
-
-    def open_or_escalate_drift(
-        self,
-        *,
-        kind: Any,
-        task_id: str,
-        agent_id: str,
-        turn_id: str,
-        severity: Any,
-    ) -> _ostate.Drift:
-        """Open a new condition or escalate an existing one (goldfive#271 PR1).
-
-        Routes through :func:`orchestration_state.open_or_escalate_drift`.
-        Same-turn re-emits for the same kind+task+agent collapse onto
-        the same condition_id; severity bumps are monotonic. The
-        returned :class:`Drift` carries the lifecycle / prev_severity
-        the caller stamps onto ``DriftDetected``.
-        """
-        if not isinstance(self._state, MutableMapping):
-            # No mutable state — fall back to a synthetic single-shot
-            # condition so the steerer's wire path still has stable
-            # condition_id / lifecycle to stamp.
-            return _ostate.Drift(
-                condition_id=_ostate.compute_condition_id(
-                    kind=kind,
-                    task_id=task_id,
-                    agent_id=agent_id,
-                    turn_id=turn_id,
-                ),
-                kind=kind,
-                task_id=str(task_id or ""),
-                agent_id=str(agent_id or ""),
-                turn_id=str(turn_id or ""),
-                severity=severity,
-                prev_severity=None,
-                lifecycle=_ostate.LIFECYCLE_OPENED,
-                occurrences=1,
-            )
-        return _ostate.open_or_escalate_drift(
-            self._state,
-            kind=kind,
-            task_id=task_id,
-            agent_id=agent_id,
-            turn_id=turn_id,
-            severity=severity,
-        )
-
-    def resolve_drift(self, condition_id: str) -> _ostate.Drift | None:
-        """Mark a condition resolved + remove it from the active set."""
-        if not isinstance(self._state, MutableMapping):
-            return None
-        return _ostate.resolve_drift(self._state, condition_id)
-
-    def escalate_to_human_intervention(self, condition_id: str) -> _ostate.Drift | None:
-        """Mark a condition as escalated to human intervention."""
-        if not isinstance(self._state, MutableMapping):
-            return None
-        return _ostate.escalate_drift_to_human_intervention(self._state, condition_id)
