+++ goldfive/runner.py
@@ -50,7 +50,7 @@ from pathlib import Path as _Path
 from typing import TYPE_CHECKING, Any
 
 from goldfive import _state_audit
-from goldfive import orchestration_state as _ostate
+from goldfive import state_store as _ostate
 from goldfive._llm import maybe_close_call_llm
 from goldfive.conversation import Conversation
 from goldfive.events import (
