+++ tests/test_orchestration_state.py
@@ -25,7 +25,7 @@ pytestmark = pytest.mark.skipif(
     reason="goldfive protobuf stubs not available (install the `dev` extra)",
 )
 
-from goldfive import orchestration_state as _ostate  # noqa: E402
+from goldfive import state_store as _ostate  # noqa: E402
 from goldfive.control import ControlKind, ControlMessage  # noqa: E402
 from goldfive.executors.sequential import SequentialExecutor  # noqa: E402
 from goldfive.reconciler import PlanReconciler  # noqa: E402
